Patch from Sebastien Vasey for permissions and symlinks issues.
Fixes a permission issue that, for example would let a package created
by checkinstall change the owner of a user's home directory to root. This is
done by ensuring permissions are preserved at each step.

An issue with symlinks is also fixed. The patch makes checkinstall
dereference all symbolic links that are not created by the installation script,
while the others are kept.